docker-dataplan-openstack-exporter/requirements.txt
prometheus_client==0.0.13
python-openstackclient==3.15.0
python-keystoneclient==3.16.0
python-novaclient==10.2.0
python-cinderclient==3.5.0
python-neutronclient==4.1.1